+++ instrumentation/opentelemetry-instrumentation-jinja2/src/opentelemetry/instrumentation/jinja2/package.py
@@ -13,4 +13,4 @@
 # limitations under the License.
 
 
-_instruments = ("jinja2~=2.7",)
+_instruments = ("jinja2 >= 2.7, < 4.0",)