+++ instrumentation/opentelemetry-instrumentation-jinja2/tests/test_jinja2.py
@@ -16,6 +16,7 @@ import os
 from unittest import mock
 
 import jinja2
+from packaging import version
 
 from opentelemetry import trace as trace_api
 from opentelemetry.instrumentation.jinja2 import Jinja2Instrumentor
@@ -31,8 +32,13 @@ class TestJinja2Instrumentor(TestBase):
         super().setUp()
         Jinja2Instrumentor().instrument()
         # prevent cache effects when using Template('code...')
-        # pylint: disable=protected-access
-        jinja2.environment._spontaneous_environments.clear()
+        if version.parse(jinja2.__version__) >= version.parse("3.0.0"):
+            # by clearing functools.lru_cache
+            jinja2.environment.get_spontaneous_environment.clear()
+        else:
+            # by clearing jinja2.utils.LRUCache
+            jinja2.environment._spontaneous_environments.clear()  # pylint: disable=no-member
+
         self.tracer = get_tracer(__name__)
 
     def tearDown(self):